requirements.txt
langchain>=0.1
langgraph>=0.0.35
openai
langchain_openai
langchainhub
langchain_experimental
langsmith
python-dotenv
fastapi
uvicorn
pymysql
openai-whisper
python-multipart
pydantic
httpx==0.24.1
psycopg2-binary
langchain_community
dnspython
requests
aiohttp
jinja2
redis 
smtplib